refactor: 파일 확장자명 변경
